pkg/logos: consolidate into one package
