+++ pkg/logos/logos.go
@@ -4,6 +4,10 @@
 
 package logos
 
+import (
+	"os"
+)
+
 type Grabber interface {
 	Lookup(name string) (*Logo, error)
 }
@@ -11,4 +15,17 @@ type Grabber interface {
 type Logo struct {
 	Name string `json:"name"`
 	URL  string `json:"url"`
+}
+
+func NewGrabber() Grabber {
+	if client := newClearbit(os.Getenv("CLEARBIT_API_KEY")); client != nil {
+		return client
+	}
+	return &noopGrabber{}
+}
+
+type noopGrabber struct{}
+
+func (g *noopGrabber) Lookup(name string) (*Logo, error) {
+	return nil, nil
 }
